refactor(pinctrl): 重构引脚控制架构，统一 API 并简化实现
- 合并 Pinctrl 和 PinManager 为统一的 PinCtrl 结构
- 删除冗余的 gpio 和 pinmanager 模块（-693 行）
- 在通用层添加 PinCtrlOp trait 和 PinCtrl 枚举，提供跨芯片抽象
- 重构 GPIO Bank 管理，集成到 PinCtrl 中
- 使用 enum_dispatch 优化枚举分发性能
- 更新 CLAUDE.md 反映新架构设计

此次重构遵循 SOLID 原则，通过统一的 trait 抽象实现了更好的可扩展性，
同时大幅简化了代码实现，提升了类型安全和可维护性。
