+++ src/variants/rk3588/mod.rs
@@ -1,7 +1,5 @@
 pub mod cru;
-pub mod gpio;
 pub mod pinctrl;
-pub mod pinmanager;
 mod syscon;
 
 // =============================================================================
@@ -9,6 +7,4 @@ mod syscon;
 // =============================================================================
 
 pub use cru::Cru;
-pub use gpio::GpioBank;
-pub use pinctrl::Pinctrl;
-pub use pinmanager::PinManager;
+pub use pinctrl::PinCtrl;
